tank-server/requirements.txt
websockets~=10.0
sanic~=22.3.2
python-socketio[asyncio_client]~=5.6.0
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
scipy>=1.4.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
tqdm>=4.41.0
pandas>=1.1.4
seaborn>=0.11.0
